+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to `laravel-support-bubble` will be documented in this file.
 
+## 1.2.0 - 2021-09-07
+
+- add container classes to config
+
+## 1.1.0 - 2021-09-07
+
+- add RTL Support (#14)
+
+## 1.0.3 - 2021-09-06
+
+- fix pointer events on hidden support form
+- don't show missing data in notifications (#11)
+
+## 1.0.2 - 2021-09-03
+
+- make submit button translatable
+
 ## 1.0.1 - 2021-09-02
 
 - replace impersonation with reply-to